+++ lib/metrics/provider.rb
@@ -35,8 +35,6 @@ module Metrics
 			
 			return provider
 		end
-		
-		Config::DEFAULT.prepare
 	else
 		def self.Provider(klass, &block)
 			# Metrics disabled.